changing the type of the price from float32 back to decimal.Decimal

+++ CandyShop/pkg/storage/postgresql/storage-postgresql.go
@@ -48,7 +48,7 @@ CREATE TABLE IF NOT EXISTS candies (
 candy_id uuid,
 category text,
 name text,
-price float,
+price numeric(18,8),
 PRIMARY KEY (candy_id, category)
 );
